+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [BREAKING] validateAndTrow is renamed to validateAndThrow (Fix #31)
 - Updated webpack configuration in Azure Functions generator to use Spawn plugin instead of shell (Fix #35)
 - principalFromBasicAuthorizationHeader is more strict and validates the presence of Basic header value (instead of being permissive) (Fix #37)
+- principalFromBearerToken is more strict and validates the presence of Bearer header value (instead of being permissive) (Fix #37)
 
 ## [0.45.0] - 2018-07-23
 ### Added