+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated webpack configuration in Azure Functions generator to use Spawn plugin instead of shell (Fix #35)
 - principalFromBasicAuthorizationHeader is more strict and validates the presence of Basic header value (instead of being permissive) (Fix #37)
 - principalFromBearerToken is more strict and validates the presence of Bearer header value (instead of being permissive) (Fix #37)
+- validationError copies all the sub-error properties now (Fix #40)
 
 ## [0.45.0] - 2018-07-23
 ### Added