+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.47.0] - 2018-08-13
 ### Added
 - uno-serverless-cli with composable commands
 - CLI: generate-schemas to generate JSON, YAML and Typescript schemas from Typescript interfaces.